Translate more tests to CXXTests
