+++ test/zcm/TransTest.hpp
@@ -1,13 +1,17 @@
-#include <iostream>
+#ifndef TRANSTEST_H
+#define TRANSTEST_H
+
 #include <thread>
-#include <cassert>
-#include <cstdio>
-#include <cstring>
-#include <signal.h>
 #include <unistd.h>
+#include <signal.h>
+
+
+#include "cxxtest/TestSuite.h"
 #include "zcm/transport_registrar.h"
 #include "util/TimeUtil.hpp"
+
 using namespace std;
+
 
 #define HZ 500
 #define MSG_COUNT 100
@@ -20,16 +24,10 @@ static void sighandler(int sig) { running_recv = false; }
 static zcm_trans_t *makeTransport()
 {
     const char *url = getenv("ZCM_DEFAULT_URL");
-    if (!url) {
-        fprintf(stderr, "ERR: Unable to find environment variable ZCM_DEFAULT_URL\n");
-        return NULL;
-    }
+    TSM_ASSERT("Unable to find environment variable ZCM_DEFAULT_URL", url);
     auto *u = zcm_url_create(url);
     auto *creator = zcm_transport_find(zcm_url_protocol(u));
-    if (!creator) {
-        fprintf(stderr, "ERR: Failed to create transport for '%s'\n", url);
-        return NULL;
-    }
+    TSM_ASSERT("Failed to create transport", creator);
     return creator(u);
 }
 
@@ -46,30 +44,23 @@ static zcm_msg_t makeMasterMsg()
     return msg;
 }
 
-#define fail(...) \
-    do {\
-        fprintf(stderr, "Err:"); \
-        fprintf(stderr, __VA_ARGS__); \
-        fprintf(stderr, "\n"); \
-        exit(1); \
-    } while(0)
-
 static void verifySame(zcm_msg_t *a, zcm_msg_t *b)
 {
     if (b->utime < a->utime || a->utime == 0 || b->utime == 0)
-        fail("Utime failure!");
+        TS_FAIL("Ultimate failure!");
     if (strcmp(a->channel, b->channel) != 0)
-        fail("Channels don't match!");
+        TS_FAIL("Channels don't match!");
     if (a->len != b->len)
-        fail("Lengths don't match!");
+        TS_FAIL("Lengths don't match!");
     for (size_t i = 0; i < a->len; i++)
         if (a->buf[i] != b->buf[i])
-            fail("Data doesn't match at index %d", (int)i);
+            TS_FAIL("Data doesn't match");
 }
 
 static void send()
 {
     auto *trans = makeTransport();
+    TSM_ASSERT("Failed to create transport", trans);
     if (!trans)
         exit(1);
 
@@ -85,6 +76,7 @@ static void send()
 static void recv()
 {
     auto *trans = makeTransport();
+    TSM_ASSERT("Failed to create transport", trans);
     if (!trans)
         exit(1);
 
@@ -92,7 +84,7 @@ static void recv()
     zcm_trans_recvmsg_enable(trans, NULL, true);
 
     zcm_msg_t master = makeMasterMsg();
-    uint64_t start = TimeUtil::utime();
+//    uint64_t start = TimeUtil::utime();
     int i;
     for (i = 0; i < MSG_COUNT && running_recv; i++) {
         zcm_msg_t msg;
@@ -101,29 +93,30 @@ static void recv()
             verifySame(&master, &msg);
         }
     }
-    uint64_t end = TimeUtil::utime();
+//    uint64_t end = TimeUtil::utime();
 
-    cout << "Received " << (i * 100 / MSG_COUNT) << "\% of the messages in "
-         << ((end - start) / 1e6) << " seconds" <<  endl;
+//    cout << "Received " << (i * 100 / MSG_COUNT) << "\% of the messages in "
+//         << ((end - start) / 1e6) << " seconds" <<  endl;
 
     running_send = false;
 }
 
-int main(int argc, char *argv[])
+class TransTest : public CxxTest::TestSuite
 {
-    if (argc > 1) {
-        string opt {argv[1]};
-        if (opt == "--small")
-            BIG_MESSAGE = false;
+  public:
+    void setUp() override {}
+    void tearDown() override {}
+
+    void testTransport() {
+        signal(SIGINT, sighandler);
+
+        std::thread sendThread {send};
+        std::thread recvThread {recv};
+
+        recvThread.join();
+        sendThread.join();
     }
+};
 
-    signal(SIGINT, sighandler);
 
-    std::thread sendThread {send};
-    std::thread recvThread {recv};
-
-    recvThread.join();
-    sendThread.join();
-
-    return 0;
-}
+#endif // TRANSTEST_H
